requirements.txt
# Core
python-binance==1.0.19
python-dotenv==1.0.0

# Data & ML
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
joblib==1.3.2

# Visualization & Dashboard
streamlit==1.32.0
plotly==5.19.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Optional: news sentiment (Bonus feature)
# requests==2.31.0
# transformers==4.37.0
# torch==2.2.0
